requirements/test.txt
-r prod.txt
coverage
pytest==7.4.3
pytest-cov
pytest-flask
tox
pyyaml==5.3.1
docker-compose
pylint==3.0.2